Extract topBar as component
