+++ frontend/src/components/sideBar.jsx
@@ -1,0 +1,5 @@
+const sideBar = () => {
+  return <></>;
+};
+
+export default sideBar;
